backend/requirements.txt
annotated-doc==0.0.4
numpy>=1.24
pandas>=2.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.5
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
isodate==0.7.2
lxml==6.0.2
more-itertools==10.8.0
msal==1.35.1
platformdirs==4.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
simple-salesforce==1.12.9
starlette==0.49.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
zeep==4.3.2
